\name{dbs-package}
\alias{dbs-package}
\alias{dbs}
\docType{package}
\title{
Accessory functions to support DBSolveOptimum
}
\description{
The package provides the extended facilities for DBSolveOptimum users. It has two main purposes: (1) the creation and analysis of DBsolveOptimum inputs and outputs for multiple simulations and statistical analysis, (2) the import/export DBSolveOptimum model and data files from/to different formats. 
}
\details{
\tabular{ll}{
Package: \tab dbs\cr
Type: \tab Package\cr
Version: \tab 0.9.1\cr
Date: \tab 2017-11-20\cr
License: \tab MIT\cr
}
DBSolveOptimum is a stand-alone software tool for construction and analysis of mathematical models of biological systems. DBSolveOptimum is implemented with new tools for extended data analysis and multiple simulations, which are important for simulation of virtual clinical trials and application of modern modeling techniques, like quantitative systems pharmacology, to problems arising in drug research and development. 

DBSolveOptimum is free for academic and industrial use. The latest version of DBSolveOptimum can be downloaded from: \url{http://sourceforge.net/projects/dbsolve/}
}
\author{
Evgeny Metelkin, Aleksey Alekseev, Aleksey Kalinkin
}
\section{Copyright }{
InSysBio, LLC\cr \url{insysbio.com}}
\references{
N.Gizzatkulov, I.Goryanin, E.Metelkin, E.Mogilevskaya, K.Peskov and O.Demin. DBSolve Optimum: a software package for kinetic modeling which allows dynamic visualization of simulation results. BMC Systems Biology, 2010, 4 (109): 1-11. \href{http://www.ncbi.nlm.nih.gov/pubmed/20698988}{PubMed}

E.Metelkin, A.Alekseev, G.Lebedeva, O.Demin. DBSolve Optimum r.33: Practical Guide \href{http://sourceforge.net/projects/dbsolve/files/User manual/DBSolve_33_manual_EN.pdf}{pdf}
}

\keyword{ package }
\seealso{
  \code{\link{calccb}}

  \code{\link{calcop}}

  \code{\link{parsetgen}}

  \code{\link{import.slv}}

  \code{\link{rct.from.slv}}

  \code{\link{import.dat}}

  \code{\link{C.from.slv}}
}
